fix(NODE-5106): refactor testing
